Feature: Add host option
The current Swagger v2.0 only supports a single host but the deployment
will likely be many different locations. Therefore having to change the
API specification for each deployment is a nightmare when dealing with
version control.

Adds ability to specify the host for the location where the APIs will
be hosted.