Feature: Switch to exclusion based options
Changes the options to be exclusion based instead of inclusion based.
This means everything is enabled by default, and the user would exclude
any feature they do not want to use. This means minimal options
required for most users. And for the corner cases the user can disable
what they do not want.